requirements.txt
certifi==2016.2.28
Django==2.0.2
djangorestframework==3.7.7
pytz==2018.3
wincertstore==0.2
